More k6 renames
