+++ tools/k6/src/web3/test/opcodes/opcodesAllEnabled.js
